pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = ""
authors = ["Nikitalskv2 <nikitka.jack@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
sqlalchemy = "^2.0.36"
pre-commit = "^4.0.1"
pydantic-settings = "^2.6.1"
bcrypt = "^4.2.0"
python-multipart = "^0.0.17"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
tzdata = "^2024.2"
asyncpg = "^0.30.0"
redis = "^5.2.0"
boto3 = "^1.35.63"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
